+++ src/seismometer/api/reports.py
@@ -3,11 +3,11 @@ from typing import Optional
 
 from seismometer.core.decorators import export
 from seismometer.data.filter import FilterRule
+from seismometer.plot.mpl.categorical import ExploreCategoricalPlots
+from seismometer.plot.mpl.categorical_single_column import ExploreSingleCategoricalPlots
 from seismometer.report.profiling import ComparisonReportWrapper, SingleReportWrapper
 from seismometer.seismogram import Seismogram
 from seismometer.table.analytics_table import ExploreBinaryModelAnalytics
-from seismometer.table.categorical import ExploreCategoricalPlots
-from seismometer.table.categorical_single_column import ExploreSingleCategoricalPlots
 from seismometer.table.fairness import ExploreBinaryModelFairness
 
 logger = logging.getLogger("seismometer")
@@ -48,11 +48,11 @@ class ExploreOrdinalMetrics(ExploreCategoricalPlots):
     Exploration widget for ordinal categorical metrics, summarizing multiple metrics for a model.
     """
 
-    def __init__(self, group_key=None):
+    def __init__(self, group_key=None, title=None):
         """
         Passes the plot function to the superclass.
         """
-        super().__init__(group_key)
+        super().__init__(group_key, title)
 
 
 @export
@@ -61,11 +61,11 @@ class ExploreCohortOrdinalMetrics(ExploreSingleCategoricalPlots):
     Exploration widget for ordinal categorical metrics, summarizing an ordinal metric across cohort subclasses.
     """
 
-    def __init__(self, group_key=None):
+    def __init__(self, group_key=None, title=None):
         """
         Passes the plot function to the superclass.
         """
-        super().__init__(group_key)
+        super().__init__(group_key, title)
 
 
 # region Reports
