+++ src/seismometer/api/utils.py
@@ -1,11 +1,14 @@
 import urllib.request
 from collections import namedtuple
 from pathlib import Path
+from typing import Optional
 
 from seismometer.configuration.config_helpers import generate_dictionary_from_parquet
 
 
-def download_example_dataset(dataset_name: str, branch_name: str = "main"):  # pragma: no cover
+def download_example_dataset(
+    dataset_name: str, branch_name: str = "main", source_repo: Optional[str] = None
+):  # pragma: no cover
     """
     Downloads an example dataset from the specified branch to local data/ directory.
 
@@ -15,6 +18,8 @@ def download_example_dataset(dataset_name: str, branch_name: str = "main"):  # p
             The name of the dataset to download.
         branch_name : str, optional
             The branch from which to download the dataset. Defaults to "main".
+        source_repo: Optional[str], optional
+            The GitHub repository that contains the data, by default None.
 
     Raises
     ------
@@ -45,7 +50,7 @@ def download_example_dataset(dataset_name: str, branch_name: str = "main"):  # p
     if dataset_name not in datasets:
         raise ValueError(f"Dataset {dataset_name} is not available in the example datasets.")
 
-    SOURCE_REPO = "epic-open-source/seismometer-data"
+    SOURCE_REPO = source_repo or "epic-open-source/seismometer-data"
     DATASET_SOURCE = f"https://raw.githubusercontent.com/{SOURCE_REPO}/refs/heads/{branch_name}/{dataset_name}"
 
     Path("data").mkdir(parents=True, exist_ok=True)
